02-data-collection/requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
alpaca-py==0.6.1
alpaca-trade-api==2.3.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
dateparser==1.1.4
deprecation==2.1.0
finnhub-python==2.4.15
frozenlist==1.3.3
GoogleNews==1.6.5
idna==3.4
kafka-python==2.0.2
lxml==4.9.2
msgpack==1.0.3
multidict==6.0.3
numpy==1.24.0
packaging==22.0
pandas==1.5.2
polygon-api-client==1.5.0
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytrends==4.8.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sseclient-py==1.7.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
websocket-client==1.4.2
websockets==10.4
yarl==1.8.2
